Fix "month offset by one"
The window is now exactly 37 months (3 years + 1 month), previously was
variable. For a few months from now this will cause a blank space at the
left of the plot, but that will go eventually. This was done because the
present issue was partially caused by related problem and a previous fix,
which is hereby reverted (27478f05b8c6a9a7c2769b7a7f0d0ad24afc7425).

QubesOS/qubes-issues#3841